+++ CHANGELOG.md
@@ -13,6 +13,12 @@
 >Any money or profit from my SDX and I will not rebrand  
 >My SDX, this decision is final and wont be changed.  
 
+>The Black arms SDX 1.11  
+>Rewritten a few things  
+>Preparing for FULL REWRITE of the SDX  
+>Added a few new functions  
+>Adding AntiYoink to this version soon  
+
 >The Black Arms SDX 1.10.10  
 >Slight expansion to Imports Pane